+++ hsg/sae/protocol/epoch.py
@@ -1,14 +1,31 @@
 import torch
 
-def train():
+def train(model, data: torch.Tensor, optimizer, loss_fn, l1_penalty) -> float:
     """
-    Executes the training stage of a single epoch.
+    Executes training optimization step for a given batch
     """
-    pass
+    model.train()
+    optimizer.zero_grad()
 
-def validate():
+    logits, latents = model(data, output_features=True)
+
+    loss = loss_fn(x=data, x_hat=logits, latents=latents, current_l1_penalty=l1_penalty)
+    loss.backward()
+
+    optimizer.step()
+
+    return loss
+
+
+def validate(model, data: torch.Tensor, loss_fn, l1_penalty) -> float:
     """
-    Executes the validation stage of a single epoch.
+    Runs the model in eval mode and returns the loss
     """
-    pass
+    with torch.no_grad():
+        model.eval()
+
+        logits, latents = model(data, output_features=True)
+        loss = loss_fn(x=data, x_hat=logits, latents=latents, current_l1_penalty=l1_penalty)
+
+        return loss
